Payment/Order One-to-One association

+++ src/main/java/com/educandoweb/course/entities/Product.java
@@ -25,7 +25,7 @@ public class Product implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private String description;
 	private Double price;
@@ -43,7 +43,7 @@ public class Product implements Serializable {
 	public Product() {}
 
 	//Não se add coleções no construtor, pois a mesma já foi instanciada lá em cima
-	public Product(Integer id, String name, String description, Double price, String imgUrl) {
+	public Product(Long id, String name, String description, Double price, String imgUrl) {
 		this.id = id;
 		this.name = name;
 		this.description = description;
@@ -51,11 +51,11 @@ public class Product implements Serializable {
 		this.imgUrl = imgUrl;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
